requirements/requirements.txt
BeautifulSoup==3.2.1
Django==1.*
Pillow==2.7.0
Pygments==2.0.2
django-bootstrap3
django-markdown-deux==1.0.5
django-xmlrpc==0.1.5
html2text==2014.12.29
markdown2==2.3.5
pytz==2014.10
textstat==0.1.6
wsgiref==0.1.2
django-xblog
django-storages
django-haystack
python-social-auth[django]
django-storages
Whoosh
